Define listdir in posix and import it in os.
My motivation is to fix https://github.com/google/pytype/issues/371.

os.listdir is in fact defined in posix: https://github.com/python/cpython/blob/9c42f8cda552694f3b47d6388d4ae84d61731872/Modules/posixmodule.c#L3826.

In Python 2, listdir is defined in both pyi files, which avoids the
linked bug but is still suboptimal, so I cleaned that up as well.
