+++ stdlib/2/os/__init__.pyi
@@ -3,7 +3,7 @@
 
 from builtins import OSError as error
 from io import TextIOWrapper as _TextIOWrapper
-from posix import stat_result as stat_result  # TODO: use this, see https://github.com/python/mypy/issues/3078
+from posix import listdir as listdir, stat_result as stat_result  # TODO: use this, see https://github.com/python/mypy/issues/3078
 import sys
 from typing import (
     Mapping, MutableMapping, Dict, List, Any, Tuple, Iterator, overload, Union, AnyStr,
@@ -200,7 +200,6 @@ def getcwd() -> str: ...
 def getcwdu() -> unicode: ...
 def chmod(path: _PathType, mode: int) -> None: ...
 def link(src: _PathType, link_name: _PathType) -> None: ...
-def listdir(path: AnyStr) -> List[AnyStr]: ...
 def lstat(path: _PathType) -> Any: ...
 def mknod(filename: _PathType, mode: int = ..., device: int = ...) -> None: ...
 def major(device: int) -> int: ...
