+++ stdlib/2/posix.pyi
@@ -1,4 +1,4 @@
-from typing import Dict, List, Mapping, Tuple, Union, Sequence, IO, Optional, TypeVar
+from typing import AnyStr, Dict, List, Mapping, Tuple, Union, Sequence, IO, Optional, TypeVar
 
 error = OSError
 
@@ -6,6 +6,8 @@ confstr_names: Dict[str, int]
 environ: Dict[str, str]
 pathconf_names: Dict[str, int]
 sysconf_names: Dict[str, int]
+
+_T = TypeVar("_T")
 
 EX_CANTCREAT: int
 EX_CONFIG: int
@@ -143,8 +145,7 @@ def kill(pid: int, sig: int) -> None: ...
 def killpg(pgid: int, sig: int) -> None: ...
 def lchown(path: unicode, uid: int, gid: int) -> None: ...
 def link(source: unicode, link_name: str) -> None: ...
-_T = TypeVar("_T")
-def listdir(path: _T) -> List[_T]: ...
+def listdir(path: AnyStr) -> List[AnyStr]: ...
 def lseek(fd: int, pos: int, how: int) -> None: ...
 def lstat(path: unicode) -> stat_result: ...
 def major(device: int) -> int: ...
