+++ stdlib/3/os/__init__.pyi
@@ -2,7 +2,7 @@
 # Ron Murawski <ron@horizonchess.com>
 
 from io import TextIOWrapper as _TextIOWrapper
-from posix import times_result
+from posix import listdir as listdir, times_result
 import sys
 from typing import (
     Mapping, MutableMapping, Dict, List, Any, Tuple, IO, Iterable, Iterator, NoReturn, overload, Union, AnyStr,
@@ -433,23 +433,6 @@ def link(
     dst_dir_fd: Optional[int] = ...,
     follow_symlinks: bool = ...,
 ) -> None: ...
-
-if sys.version_info >= (3, 6):
-    @overload
-    def listdir(path: Optional[str] = ...) -> List[str]: ...
-    @overload
-    def listdir(path: bytes) -> List[bytes]: ...
-    @overload
-    def listdir(path: int) -> List[str]: ...
-    @overload
-    def listdir(path: PathLike[str]) -> List[str]: ...
-else:
-    @overload
-    def listdir(path: Optional[str] = ...) -> List[str]: ...
-    @overload
-    def listdir(path: bytes) -> List[bytes]: ...
-    @overload
-    def listdir(path: int) -> List[str]: ...
 
 def lstat(path: _PathType, *, dir_fd: Optional[int] = ...) -> stat_result: ...
 def mkdir(path: _PathType, mode: int = ..., *, dir_fd: Optional[int] = ...) -> None: ...
